# spec/dashboard/games_openapi.yaml
openapi: 3.0.0
info:
  description: "Qilin api "
  version: "1.0.0"
  title: "Qilin games dashboard API"
  contact:
    email: "roman.golenok@protocol.one"
...
servers:
  - url: https://qilinapi.tst.protocol.one
    description: Master test server
paths:
...
  /api/v1/games/{id}:
    get:
      tags:
        - "game"
      summary: "Get game info"
      description: ""
      parameters:
        - name: "id"
          in: "path"
          schema:
            type: string
          required: true
          description: "Id of game"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameInfo"
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalError'


    delete:
      tags:
        - "game"
      summary: "Delete game"
      description: ""
      operationId: "deleteGame"
      parameters:
        - name: "id"
          in: "path"
          schema:
            type: string
          required: true
          description: "Id of game to remove"
      responses:
        200:
          description: OK
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalError'

    put:
      tags:
        - "game"
      summary: "Change game info"
      description: ""
      operationId: "changeGame"
      requestBody:
        description: "Game object that needs to be saved in store"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeGame"
      parameters:
        - name: "id"
          in: "path"
          schema:
            type: string
            format: uuid
          required: true
          description: "Id of game to change"
        - name: If-Unmodified-Since
          in: header
          schema:
            type: string
            format: date-time
          required: false
      responses:
        200:
          description: OK
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        409:
          $ref: '#/components/responses/Conflict'
        422:
          $ref: '#/components/responses/UnprocessableEntity'
        500:
          $ref: '#/components/responses/InternalError'
...
  /api/v1/games/{id}/discounts:
    get:
      tags:
        - price
      summary: "Gets discounts for game"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Discount'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalError'

    post:
      tags:
        - price
        - discount
      summary: "Create new discount for game"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiscountRequest'
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Discount'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        409:
          $ref: '#/components/responses/Conflict'
        422:
          $ref: '#/components/responses/UnprocessableEntity'
        500:
          $ref: '#/components/responses/InternalError'
  /api/v1/games/{id}/discounts/{discountId}:
    put:
      tags:
        - price
        - discount
      summary: "Change discount for game"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: If-Unmodified-Since
          in: header
          schema:
            type: string
            format: date-time
          required: false
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiscountRequest'
      responses:
        200:
          description: OK
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        409:
          $ref: '#/components/responses/Conflict'
        422:
          $ref: '#/components/responses/UnprocessableEntity'
        500:
          $ref: '#/components/responses/InternalError'

    delete:
      tags:
        - price
        - discount
      summary: "Remove discount for game"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: OK
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal Server Error
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 409
            message: "Resource with same unique id already exists"
    UnprocessableEntity:
      description: Unprocessable Entity
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 422
            message: "Fields required: 'title, description'"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    DiscountRequest:
      type: object
      properties:
        title:
          type: object
          additionalProperties:
            type: string
          example:
            en: "WINTER SALE"
            ru: "ЗИМНЯЯ РАСПРОДАЖА"
        description:
          type: object
          additionalProperties:
            type: string
          example:
            en: "SPECIAL OFFER"
            ru: "СПЕЦИАЛЬНОЕ ПРЕДЛОЖЕНИЕ"
        date:
          type: object
          properties:
            start:
              type: string
              format: date-time
              example: "2019-07-21T17:32:28Z"
            end:
              type: string
              format: date-time
              example: "2019-08-21T17:32:28Z"
        rate:
          type: number
          format: float
          example: 50
    Discount:
      allOf:
        - $ref: '#/components/schemas/DiscountRequest'
        - type: object
          properties:
            id:
              type: string
              format: uuid
              example: "028536a6-3e8e-4926-9669-1cd8f31d8f61"
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
...
    LanguageSupport:
      type: object
      properties:
        voice:
          type: "boolean"
          default: false
        interface:
          type: "boolean"
          default: false
        subtitles:
          type: "boolean"
          default: false
    Languages:
      type: object
      additionalProperties:
        $ref: "#/components/schemas/LanguageSupport"
      example:
        en:
          voice: true
          interface: true
          subtitles: true
        ru:
          voice: false
          interface: true
          subtitles: true
        fr:
          voice: false
          interface: false
          subtitles: true
...
    GameGenres:
      type: object
      properties:
        main:
          type: integer
        addition:
          type: array
          items:
            type: integer
...
    Features:
      type: object
      properties:
        common:
          type: array
          items:
            type: string
            enum:
              - singlePlayer
              - multiPlayer
              - onlineMultiplayerPlayer
              - localMultiplayerPlayer
              - splitScreen
              - mmo
              - cooperative
              - onlineCooperative
              - localCooperative
              - crossPlatform
              - achievements
              - cloudSaves
        controllers:
          type: "string"
          enum:
            - "not_supported"
            - "partially_supported"
            - "full_supported"
          default: "not_supported"
    SupportPlatform:
      type: object
      properties:
        windows:
          type: "boolean"
          default: true
        macOs:
          type: "boolean"
          default: false
        linux:
          type: "boolean"
          default: false
    MachineRequirements:
      type: object
      properties:
        system:
          type: "string"
          example: "Windows"
        processor:
          type: "string"
          example: "Intel i9"
        graphics:
          type: "string"
          example: "NVIDIA GeForce RTX2080Ti or similar"
        sound:
          type: "string"
        ram:
          type: "integer"
          example: 4
        ramDimension:
          type: string
          enum:
            - MB
            - GB
        storage:
          type: "integer"
        storageDimension:
          type: string
          enum:
            - MB
            - GB
        other:
          type: "string"
...
    BaseGame:
      type: object
      properties:
        internalName:
          type: "string"
          readOnly: true
        # non-localized title of game for end user
        title:
          type: "string"
        # comma separated list of developers
        developers:
          type: "string"
        # comma separated list of publishers
        publishers:
          type: "string"
        releaseDate:
          type: "string"
          format: "date-time"
        displayRemainingTime:
          type: "boolean"
          default: false
        features:
          $ref: "#/components/schemas/Features"
        platforms:
          $ref: "#/components/schemas/SupportPlatform"
        requirements:
          type: object
          properties:
            macOs:
              type: object
              properties:
                minimal:
                  $ref: "#/components/schemas/MachineRequirements"
                recommended:
                  $ref: "#/components/schemas/MachineRequirements"
            linux:
              type: object
              properties:
                minimal:
                  $ref: "#/components/schemas/MachineRequirements"
                recommended:
                  $ref: "#/components/schemas/MachineRequirements"
            windows:
              type: object
              properties:
                minimal:
                  $ref: "#/components/schemas/MachineRequirements"
                recommended:
                  $ref: "#/components/schemas/MachineRequirements"
        languages:
          $ref: "#/components/schemas/Languages"
        achievementOnProd:
          type: boolean
          default: false
          description: "Enables achievements on production"
    GameInfo:
      allOf:
        - $ref: '#/components/schemas/BaseGame'
        - type: object
          properties:
            genres:
              $ref: '#/components/schemas/GameGenres'
            tags:
              type: array
              items:
                type: integer
    ChangeGame:
      allOf:
        - $ref: '#/components/schemas/BaseGame'
        - type: object
          properties:
            genres:
              $ref: '#/components/schemas/GameGenres'
            tags:
              type: array
              items:
                type: integer
...
security:
  - bearerAuth: []